+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 - Add `.with_defaults()` as a factory builder for terminals with custom defaults.
+- Add `default_action` setting to configure what the picker on selecting with `<Enter>`.
 
 ## 0.4.0 - 2025-09-22